Updated PowerShell code
I think the PowerShell Code on this page is not easily understandable and it did not work in my environment. Why do we need this line ?
$PoolCount = Get-PhysicalDisk -CanPool $True

$PoolCount is not used anywhere onwards

This line does not give me any Data back
$PhysicalDisks = Get-PhysicalDisk | Where-Object {$_.FriendlyName -like "*2" -or $_.FriendlyName -like "*3"}

The rest is a really messy pipeline ;). 

I think my code change can be more easily understood.